+++ methods/L2P_old.py
@@ -35,8 +35,6 @@ from torch.utils.data import DataLoader
 from torch.utils.tensorboard import SummaryWriter
 from torch import optim
 
-from methods._trainer import _Trainer
-
 from utils.data_loader import ImageDataset, StreamDataset, MemoryDataset, cutmix_data, get_statistics
 from utils.train_utils import select_model, select_optimizer, select_scheduler
 
@@ -159,13 +157,12 @@ class L2P_Model(nn.Module):
         self.add_module('backbone', timm.models.create_model(backbone_name, pretrained=True, num_classes=class_num,
                                                              drop_rate=0.,drop_path_rate=0.,drop_block_rate=None))
         for name, param in self.backbone.named_parameters():
-            # print(name)
             if 'head' not in name:
                 param.requires_grad = False
-        self.backbone.fc.weight.requires_grad = True
-        self.backbone.fc.bias.requires_grad   = True
+        self.backbone.head.weight.requires_grad = True
+        self.backbone.head.bias.requires_grad   = True
 
-        # self.head = self.backbone.fc
+        self.head = self.backbone.head
         
         self.prompt = Prompt(
             pool_size,
@@ -197,7 +194,7 @@ class L2P_Model(nn.Module):
         x = self.backbone.norm(x)
         x = x[:, 1:self.selection_size * self.prompt_len + 1].clone()
         x = x.mean(dim=1)
-        x = self.backbone.fc(x)
+        x = self.backbone.head(x)
         return x
     
     def loss_fn(self, output, target):
@@ -219,149 +216,127 @@ class L2P_Model(nn.Module):
     
 
 
-class L2P(_Trainer):
-    def __init__(self,*args,**kwargs):
-        super(L2P,self).__init__(*args,**kwargs)
-    # def __init__(
-    #         self, criterion, device, train_transform, test_transform, n_classes, **kwargs
-    # ):
-        # self.num_learned_class = 0
-        # self.num_learning_class = 1
-        # self.n_classes = kwargs["n_classes"]
-        # self.exposed_classes = []
-        # self.seen = 0
-        # self.topk = kwargs["topk"]
+class L2P(ER):
+    def __init__(
+            self, criterion, device, train_transform, test_transform, n_classes, **kwargs
+    ):
+        self.num_learned_class = 0
+        self.num_learning_class = 1
+        self.n_classes = n_classes
+        self.exposed_classes = []
+        self.seen = 0
+        self.topk = kwargs["topk"]
 
-        # self.device = kwargs["device"]
-        # self.dataset = kwargs["dataset"]
-        # self.model_name = kwargs["model_name"]
-        # self.opt_name = kwargs["opt_name"]
-        # self.sched_name = kwargs["sched_name"]
-        # if self.sched_name == "default":
-        #     self.sched_name = 'exp_reset'
-        # self.lr = kwargs["lr"]
+        self.device = device
+        self.dataset = kwargs["dataset"]
+        self.model_name = kwargs["model_name"]
+        self.opt_name = kwargs["opt_name"]
+        self.sched_name = kwargs["sched_name"]
+        if self.sched_name == "default":
+            self.sched_name = 'exp_reset'
+        self.lr = kwargs["lr"]
 
-        # self.train_transform = kwargs["train_transform"]
-        # self.cutmix = "cutmix" in kwargs["transforms"]
-        # self.test_transform = kwargs["test_transform"]
+        self.train_transform = train_transform
+        self.cutmix = "cutmix" in kwargs["transforms"]
+        self.test_transform = test_transform
 
-        # self.memory_size = kwargs["memory_size"]
-        # self.data_dir = kwargs["data_dir"]
+        self.memory_size = kwargs["memory_size"]
+        self.data_dir = kwargs["data_dir"]
 
-        # self.online_iter = kwargs["online_iter"]
-        # self.batch_size = kwargs["batchsize"]
+        self.online_iter = kwargs["online_iter"]
+        self.batch_size = kwargs["batchsize"]
         
-        # self.temp_batchsize = kwargs["temp_batchsize"]
-        # if self.temp_batchsize is None:
-        #     self.temp_batchsize = self.batch_size//2
-        # if self.temp_batchsize > self.batch_size:
-        #     self.temp_batchsize = self.batch_size
+        self.temp_batchsize = kwargs["temp_batchsize"]
+        if self.temp_batchsize is None:
+            self.temp_batchsize = self.batch_size//2
+        if self.temp_batchsize > self.batch_size:
+            self.temp_batchsize = self.batch_size
         
-        # self.memory_size -= self.temp_batchsize
+        self.memory_size -= self.temp_batchsize
 
-        # self.gpu_transform = kwargs["gpu_transform"]
-        # self.use_amp = kwargs["use_amp"]
-        # if self.use_amp:
-        #     self.scaler = torch.cuda.amp.GradScaler()
+        self.gpu_transform = kwargs["gpu_transform"]
+        self.use_amp = kwargs["use_amp"]
+        if self.use_amp:
+            self.scaler = torch.cuda.amp.GradScaler()
 
 
-        self.model = L2P_Model(backbone_name='vit_base_patch16_224_l2p', class_num=1)
-        # for n,p in self.model.named_parameters():
-        #     print(n)
+        self.model = L2P_Model(backbone_name='vit_base_patch16_224_l2p', class_num=1).to(self.device)
+        self.criterion = self.model.loss_fn
 
-        # self.criterion = kwargs["criterion"]
+        params = [param for name, param in self.model.named_parameters() if 'head' not in name]
+        self.optimizer = select_optimizer(self.opt_name, self.lr, self.model, True)
+        if 'imagenet' in self.dataset:
+            self.lr_gamma = 0.99995
+        else:
+            self.lr_gamma = 0.9999
+        # self.optimizer.add_param_group({'params': self.model.backbone.head.parameters()})
+        self.scheduler = select_scheduler(self.sched_name, self.optimizer, self.lr_gamma)
+        # self.memory = MemoryDataset(self.train_transform, cls_list=self.exposed_classes,
+        #                             test_transform=self.test_transform)
+        self.temp_batch = []
+        self.temp_label = []
+        self.num_updates = 0
+        self.train_count = 0
+        self.batch_size = kwargs["batchsize"]
 
-        # params = [param for name, param in self.model.named_parameters() if 'head' not in name]
-        # self.optimizer = select_optimizer(self.opt_name, self.lr, self.model, True)
-        # if 'imagenet' in self.dataset:
-        #     self.lr_gamma = 0.99995
-        # else:
-        #     self.lr_gamma = 0.9999
-        # # self.optimizer.add_param_group({'params': self.model.backbone.head.parameters()})
-        # self.scheduler = select_scheduler(self.sched_name, self.optimizer, self.lr_gamma)
-        # # self.memory = MemoryDataset(self.train_transform, cls_list=self.exposed_classes,
-        # #                             test_transform=self.test_transform)
-        # self.temp_batch = []
-        # self.temp_label = []
-        # self.num_updates = 0
-        # self.train_count = 0
-        # self.batch_size = kwargs["batchsize"]
+        self.start_time = time.time()
+        num_samples = {'cifar10': 50000, 'cifar100': 50000, 'tinyimagenet': 100000, 'imagenet': 1281167}
+        self.total_samples = num_samples[self.dataset]
+        # if self.dataset=='cifar10':
+        self.convert_li = ['airplane','automobile','bird','cat','deer','dog','frog','horse','ship','truck']
 
-        # self.start_time = time.time()
-        # num_samples = {'cifar10': 50000, 'cifar100': 50000, 'tinyimagenet': 100000, 'imagenet': 1281167}
-        # self.total_samples = num_samples[self.dataset]
-        
-    #     # if self.dataset=='cifar10':
-    #     self.convert_li = ['airplane','automobile','bird','cat','deer','dog','frog','horse','ship','truck']
-
-    def online_step(self, sample, sample_num):
+    def online_step(self, sample, sample_num, n_worker):
         image, label = sample
         for l in label:
             if l.item() not in self.exposed_classes:
                 self.add_new_class(l.item())
 
-        self.num_updates += self.online_iter * self.batchsize
-        train_loss, train_acc = self.online_train([image, label], iterations=int(self.num_updates))
-        # self.report_training(sample_num, train_loss, train_acc)
+        self.num_updates += self.online_iter * self.batch_size
+        train_loss, train_acc = self.online_train([image, label], self.batch_size * 2, n_worker,
+                                                    iterations=int(self.num_updates), stream_batch_size=self.batch_size)
+        self.report_training(sample_num, train_loss, train_acc)
         # for stored_sample, stored_label in zip(image, label):
         #     self.update_memory((stored_sample, stored_label))
-        
+        self.temp_batch = []
         self.num_updates -= int(self.num_updates)
-        return train_loss, train_acc
 
-    # def add_new_class(self, class_name):
-    #     self.exposed_classes.append(class_name)
-    #     self.num_learned_class = len(self.exposed_classes)
-    #     prev_weight = copy.deepcopy(self.model_without_ddp.fc.weight.data)
-    #     prev_bias   = copy.deepcopy(self.model_without_ddp.fc.bias.data)
-    #     self.model_without_ddp.fc = nn.Linear(self.model_without_ddp.fc.in_features, self.num_learned_class).to(self.device)
-        
-    #     # self.model.backbone.reset_classifier(self.num_learned_class)
-
-    #     # self.model.backbone.head.to(self.device)
-    #     with torch.no_grad():
-    #         if self.num_learned_class > 1:
-    #             self.model_without_ddp.fc.weight[:self.num_learned_class - 1] = prev_weight
-    #             self.model_without_ddp.fc.bias[:self.num_learned_class - 1]   = prev_bias
-    #     for param in self.optimizer.param_groups[1]['params']:
-    #         if param in self.optimizer.state.keys():
-    #             del self.optimizer.state[param]
-    #     del self.optimizer.param_groups[1]
-    #     self.optimizer.add_param_group({'params': self.model_without_ddp.fc.parameters()})
-    #     # self.scheduler = select_scheduler(self.sched_name, self.optimizer, self.lr_gamma)
-    #     # self.memory.add_new_class(cls_list=self.exposed_classes)
-    #     if 'reset' in self.sched_name:
-    #         self.update_schedule(reset=True)
-            
     def add_new_class(self, class_name):
         self.exposed_classes.append(class_name)
         self.num_learned_class = len(self.exposed_classes)
-        prev_weight = copy.deepcopy(self.model_without_ddp.fc.weight.data)
-        prev_bias = copy.deepcopy(self.model_without_ddp.fc.bias.data)
-        self.model_without_ddp.fc = nn.Linear(self.model_without_ddp.fc.in_features, self.num_learned_class).to(self.device)
+        prev_weight = copy.deepcopy(self.model.backbone.head.weight.data)
+        prev_bias   = copy.deepcopy(self.model.backbone.head.bias.data)
+        self.model.backbone.reset_classifier(self.num_learned_class)
+
+        self.model.backbone.head.to(self.device)
         with torch.no_grad():
             if self.num_learned_class > 1:
-                self.model_without_ddp.fc.weight[:self.num_learned_class - 1] = prev_weight
-                self.model_without_ddp.fc.bias[:self.num_learned_class - 1]   = prev_bias
+                self.model.backbone.head.weight[:self.num_learned_class - 1] = prev_weight
+                self.model.backbone.head.bias[:self.num_learned_class - 1]   = prev_bias
         for param in self.optimizer.param_groups[1]['params']:
             if param in self.optimizer.state.keys():
                 del self.optimizer.state[param]
         del self.optimizer.param_groups[1]
-        params = [param for name, param in self.model.named_parameters() if 'fc' in name]
-        self.optimizer.add_param_group({'params': params})
+        self.optimizer.add_param_group({'params': self.model.backbone.head.parameters()})
+        self.scheduler = select_scheduler(self.sched_name, self.optimizer, self.lr_gamma)
         # self.memory.add_new_class(cls_list=self.exposed_classes)
         if 'reset' in self.sched_name:
             self.update_schedule(reset=True)
 
-    def online_train(self, sample, iterations=1):
+    def online_train(self, sample, batch_size, n_worker, iterations=1, stream_batch_size=1):
         
         total_loss, correct, num_data = 0.0, 0.0, 0.0
 
+        # if len(self.memory) > 0 and batch_size - stream_batch_size > 0:
+        #     memory_batch_size = min(len(self.memory), batch_size - stream_batch_size)
         for i in range(iterations):
             self.model.train()
             x, y = sample
             x = torch.cat([self.train_transform(transforms.ToPILImage()(img)).unsqueeze(0) for img in x])
             y = torch.cat([torch.tensor([self.exposed_classes.index(label)]) for label in y])
+            # if len(self.memory) > 0:
+            #     memory_data = self.memory.get_batch(memory_batch_size)
+            #     x = torch.cat([x, memory_data['image']])
+            #     y = torch.cat([y, memory_data['label']])
             x = x.to(self.device)
             y = y.to(self.device)
 
@@ -408,22 +383,25 @@ class L2P(_Trainer):
                 loss = self.criterion(logit, y)
         return logit, loss
 
-    # def report_training(self, sample_num, train_loss, train_acc):
-    #     writer.add_scalar(f"train/loss", train_loss, sample_num)
-    #     writer.add_scalar(f"train/acc", train_acc, sample_num)
-    #     logger.info(
-    #         f"Train | Sample # {sample_num} | train_loss {train_loss:.4f} | train_acc {train_acc:.4f} | "
-    #         f"lr {self.optimizer.param_groups[0]['lr']:.6f} | "
-    #         f"running_time {datetime.timedelta(seconds=int(time.time() - self.start_time))} | "
-    #         f"ETA {datetime.timedelta(seconds=int((time.time() - self.start_time) * (self.total_samples-sample_num) / sample_num))}"
-    #     )
+    def report_training(self, sample_num, train_loss, train_acc):
+        writer.add_scalar(f"train/loss", train_loss, sample_num)
+        writer.add_scalar(f"train/acc", train_acc, sample_num)
+        logger.info(
+            f"Train | Sample # {sample_num} | train_loss {train_loss:.4f} | train_acc {train_acc:.4f} | "
+            f"lr {self.optimizer.param_groups[0]['lr']:.6f} | "
+            f"running_time {datetime.timedelta(seconds=int(time.time() - self.start_time))} | "
+            f"ETA {datetime.timedelta(seconds=int((time.time() - self.start_time) * (self.total_samples-sample_num) / sample_num))}"
+        )
 
-    # def report_test(self, sample_num, avg_loss, avg_acc):
-    #     writer.add_scalar(f"test/loss", avg_loss, sample_num)
-    #     writer.add_scalar(f"test/acc", avg_acc, sample_num)
-    #     logger.info(
-    #         f"Test | Sample # {sample_num} | test_loss {avg_loss:.4f} | test_acc {avg_acc:.4f} | "
-    #     )
+    def report_test(self, sample_num, avg_loss, avg_acc):
+        writer.add_scalar(f"test/loss", avg_loss, sample_num)
+        writer.add_scalar(f"test/acc", avg_acc, sample_num)
+        logger.info(
+            f"Test | Sample # {sample_num} | test_loss {avg_loss:.4f} | test_acc {avg_acc:.4f} | "
+        )
+
+    def update_memory(self, sample):
+        self.reservoir_memory(sample)
 
     def update_schedule(self, reset=False):
         if reset:
@@ -433,69 +411,52 @@ class L2P(_Trainer):
         else:
             self.scheduler.step()
 
-    # def online_evaluate(self, test_loader):
-    #     eval_dict = self.evaluation(test_loader, self.criterion)
-    #     # self.report_test(sample_num, eval_dict["avg_loss"], eval_dict["avg_acc"])
-        
-    #     total_correct, total_num_data, total_loss = 0.0, 0.0, 0.0
-    #     correct_l = torch.zeros(self.n_classes)
-    #     num_data_l = torch.zeros(self.n_classes)
-    #     label = []
+    def online_evaluate(self, test_loader, sample_num):
+        eval_dict = self.evaluation(test_loader, self.criterion)
+        self.report_test(sample_num, eval_dict["avg_loss"], eval_dict["avg_acc"])
+        return eval_dict
 
-    #     self.model.eval()
-    #     with torch.no_grad():
-    #         for i, data in enumerate(test_loader):
-    #             x, y = data
-    #             for j in range(len(y)):
-    #                 y[j] = self.exposed_classes.index(y[j].item())
-    #             x = x.to(self.device)
-    #             y = y.to(self.device)
-    #             logit = self.model(x)
-
-    #             loss = self.criterion(logit, y)
-    #             pred = torch.argmax(logit, dim=-1)
-    #             _, preds = logit.topk(self.topk, 1, True, True)
-
-    #             total_correct += torch.sum(preds == y.unsqueeze(1)).item()
-    #             total_num_data += y.size(0)
-
-    #             xlabel_cnt, correct_xlabel_cnt = self._interpret_pred(y, pred)
-    #             correct_l += correct_xlabel_cnt.detach().cpu()
-    #             num_data_l += xlabel_cnt.detach().cpu()
-
-    #             total_loss += loss.item()
-    #             label += y.tolist()
-
-    #     avg_acc = total_correct / total_num_data
-    #     avg_loss = total_loss / len(test_loader)
-    #     cls_acc = (correct_l / (num_data_l + 1e-5)).numpy().tolist()
-        
-    #     eval_dict = {"avg_loss": avg_loss, "avg_acc": avg_acc, "cls_acc": cls_acc}
-
-    #     return eval_dict
-
-
-    def online_before_task(self,task_id):
+    def online_before_task(self,train_loader,debug):
         #todo 현재 Task Class 및 Sample 확인
-        # print("self.model")
-        # for n,p in self.model.named_parameters():
-        #     print(n)
+        data_info = {}
+        for i, data in enumerate(train_loader):
+            # if debug and (i+1)*self.batch_size == 200:
+            #     break
+            _,label = data
+            # image = image.to(self.device)
+            label = label.to(self.device)
+            
+            for b in range(label.shape[0]):
+                if 'Class_'+str(label[b].item()) in data_info.keys():
+                    data_info['Class_'+str(label[b].item())] +=1
+                else:
+                    data_info['Class_'+str(label[b].item())] =1
         
-        # print("self.model_without_ddp")
-        # for n,p in self.model_without_ddp.named_parameters():
-        #     print(n)
-
-        pass
+        print("Current Task Data Info")
+        print(data_info)
+        print("<<Convert to str>>")
+        convert_data_info = self.convert_class_from_int_to_str(data_info)
+        print(convert_data_info)
+        print()
 
     def online_after_task(self, cur_iter):
         # Task-Free
         pass
 
+    def reservoir_memory(self, sample):
+        self.seen += 1
+        if len(self.memory.images) >= self.memory_size:
+            j = np.random.randint(0, self.seen)
+            if j < self.memory_size:
+                self.memory.replace_sample(sample, j)
+        else:
+            self.memory.replace_sample(sample)
+
     def reset_opt(self):
         self.optimizer = select_optimizer(self.opt_name, self.lr, self.model, True)
         self.scheduler = select_scheduler(self.sched_name, self.optimizer, self.lr_gamma)
 
-    def online_evaluate(self, test_loader):
+    def evaluation(self, test_loader, criterion):
         total_correct, total_num_data, total_loss = 0.0, 0.0, 0.0
         correct_l = torch.zeros(self.n_classes)
         num_data_l = torch.zeros(self.n_classes)
@@ -549,55 +510,55 @@ class L2P(_Trainer):
         return ret_num_data, ret_corrects
     
 
-    # def train_data_config(self,n_task, train_dataset,train_sampler):
-    #     from torch.utils.data import DataLoader
-    #     for t_i in range(n_task):
-    #         train_sampler.set_task(t_i)
-    #         train_dataloader= DataLoader(train_dataset, batch_size=self.batch_size, sampler=train_sampler, num_workers=4)
-    #         data_info = {}
-    #         for i, data in enumerate(train_dataloader):
-    #             _,label = data
-    #             label = label.to(self.device)
+    def train_data_config(self,n_task, train_dataset,train_sampler):
+        from torch.utils.data import DataLoader
+        for t_i in range(n_task):
+            train_sampler.set_task(t_i)
+            train_dataloader= DataLoader(train_dataset, batch_size=self.batch_size, sampler=train_sampler, num_workers=4)
+            data_info = {}
+            for i, data in enumerate(train_dataloader):
+                _,label = data
+                label = label.to(self.device)
                 
-    #             for b in range(len(label)):
-    #                 if 'Class_'+str(label[b].item()) in data_info.keys():
-    #                     data_info['Class_'+str(label[b].item())] +=1
-    #                 else:
-    #                     data_info['Class_'+str(label[b].item())] =1
-    #         print(f"[Train] Task {t_i} Data Info")
-    #         convert_data_info = self.convert_class_from_int_to_str(data_info)
-    #         print(convert_data_info)
-    #         print()
+                for b in range(len(label)):
+                    if 'Class_'+str(label[b].item()) in data_info.keys():
+                        data_info['Class_'+str(label[b].item())] +=1
+                    else:
+                        data_info['Class_'+str(label[b].item())] =1
+            print(f"[Train] Task {t_i} Data Info")
+            convert_data_info = self.convert_class_from_int_to_str(data_info)
+            print(convert_data_info)
+            print()
     
-    # def test_data_config(self, test_dataloader,task_id):
-    #     from torch.utils.data import DataLoader
-    #     data_info = {}
-    #     for i, data in enumerate(test_dataloader):
-    #         _,label = data
-    #         label = label.to(self.device)
+    def test_data_config(self, test_dataloader,task_id):
+        from torch.utils.data import DataLoader
+        data_info = {}
+        for i, data in enumerate(test_dataloader):
+            _,label = data
+            label = label.to(self.device)
             
-    #         for b in range(len(label)):
-    #             if 'Class_'+str(label[b].item()) in data_info.keys():
-    #                 data_info['Class_'+str(label[b].item())] +=1
-    #             else:
-    #                 data_info['Class_'+str(label[b].item())] =1
+            for b in range(len(label)):
+                if 'Class_'+str(label[b].item()) in data_info.keys():
+                    data_info['Class_'+str(label[b].item())] +=1
+                else:
+                    data_info['Class_'+str(label[b].item())] =1
                     
-    #     print('<<Exposed Class>>')
-    #     print(self.exposed_classes)
+        print('<<Exposed Class>>')
+        print(self.exposed_classes)
         
-    #     print(f"[Test] Task {task_id} Data Info")
-    #     print(data_info)
-    #     print("<<Convert>>")
-    #     convert_data_info = self.convert_class_from_int_to_str(data_info)
-    #     print(convert_data_info)
-    #     print()
+        print(f"[Test] Task {task_id} Data Info")
+        print(data_info)
+        print("<<Convert>>")
+        convert_data_info = self.convert_class_from_int_to_str(data_info)
+        print(convert_data_info)
+        print()
         
         
-    # def convert_class_from_int_to_str(self,data_info):
+    def convert_class_from_int_to_str(self,data_info):
         
-    #     self.convert_li
-    #     for key in list(data_info.keys()):
-    #         old_key = int(key[6:])
-    #         data_info[self.convert_li[old_key]] = data_info.pop(key)
+        self.convert_li
+        for key in list(data_info.keys()):
+            old_key = int(key[6:])
+            data_info[self.convert_li[old_key]] = data_info.pop(key)
         
-    #     return data_info
+        return data_info
