+++ scripts/l2p_rnd_NM.sh
@@ -1,14 +1,14 @@
 #!/bin/bash
 
-#SBATCH -J L2P_iblurry_cifar100_N50_M10_RND_test
+#SBATCH -J L2P_iblurry_cifar100_N50_M10_RND_test2_seed1
 #SBATCH -p batch
-#SBATCH -w vll1
+#SBATCH -w agi2
 #SBATCH --nodes=1
-#SBATCH --gres=gpu:1
+#SBATCH --gres=gpu:2
 #SBATCH --cpus-per-gpu=4
 #SBATCH --mem-per-gpu=20G
 #SBATCH --time=14-0
-#SBATCH -o %x_%j.log
+#SBATCH -o %x_%j.out
 
 date
 ulimit -n 65536
@@ -29,38 +29,38 @@ conda activate torch38gpu
 
 conda --version
 python --version
-
+echo "Batch size 64 onlin iter 3"
 # CIL CONFIG
-NOTE="L2P_iblurry_cifar10_N50_M10_RND_test" # Short description of the experiment. (WARNING: logs/results with the same note will be overwritten!)
+NOTE="L2P_iblurry_cifar100_N50_M10_RND_test2_seed1" # Short description of the experiment. (WARNING: logs/results with the same note will be overwritten!)
 MODE="L2P"
-DATASET="cifar10" # cifar10, cifar100, tinyimagenet, imagenet
+DATASET="cifar100" # cifar10, cifar100, tinyimagenet, imagenet
 N_TASKS=5
 N=50
 M=10
 GPU_TRANSFORM="--gpu_transform"
 USE_AMP="--use_amp"
-SEEDS="1 2 3"
+SEEDS="1"
 VIT="True"
 OPT="adam"
 
 if [ "$DATASET" == "cifar10" ]; then
     MEM_SIZE=500 ONLINE_ITER=1
-    MODEL_NAME="L2P" EVAL_PERIOD=100
-    BATCHSIZE=16; LR=0.03 OPT_NAME="adam" SCHED_NAME="const" MEMORY_EPOCH=256
+    MODEL_NAME="vit" EVAL_PERIOD=1000
+    BATCHSIZE=128; LR=0.03 OPT_NAME="adam" SCHED_NAME="const" MEMORY_EPOCH=256
 
 elif [ "$DATASET" == "cifar100" ]; then
-    MEM_SIZE=2000 ONLINE_ITER=1
-    MODEL_NAME="L2P" EVAL_PERIOD=100
-    BATCHSIZE=16; LR=0.0075 OPT_NAME="adam" SCHED_NAME="const" MEMORY_EPOCH=256
+    MEM_SIZE=2000 ONLINE_ITER=3
+    MODEL_NAME="vit" EVAL_PERIOD=1000
+    BATCHSIZE=64; LR=0.0075 OPT_NAME="adam" SCHED_NAME="const" MEMORY_EPOCH=256
 
 elif [ "$DATASET" == "tinyimagenet" ]; then
     MEM_SIZE=4000 ONLINE_ITER=3
-    MODEL_NAME="L2P" EVAL_PERIOD=100
+    MODEL_NAME="vit" EVAL_PERIOD=100
     BATCHSIZE=32; LR=0.03 OPT_NAME="adam" SCHED_NAME="const" MEMORY_EPOCH=256
 
 elif [ "$DATASET" == "imagenet" ]; then
     N_TASKS=10 MEM_SIZE=20000 ONLINE_ITER=0.25
-    MODEL_NAME="L2P" EVAL_PERIOD=1000
+    MODEL_NAME="vit" EVAL_PERIOD=1000
     BATCHSIZE=256; LR=0.03 OPT_NAME="adam" SCHED_NAME="const" MEMORY_EPOCH=100
     BATCHSIZE=256; LR=0.05 OPT_NAME=$OPT SCHED_NAME="multistep" MEMORY_EPOCH=100
