--- utils/method_manager.py
@@ -1,146 +1,0 @@
-import logging
-
-from methods.bic import BiasCorrection
-from methods.er_baseline import ER
-from methods.rainbow_memory import RM
-from methods.ewc import EWCpp
-from methods.mir import MIR
-from methods.clib import CLIB
-from models.L2P import L2P
-from methods.ViT_finetuning import ViT_FT
-from methods.ViT_Linear import ViT_LP
-
-# from methods.Freeze_extractor import Freeze_extractor
-# from methods.er_baseline_ViT import ER_ViT
-# from methods.rainbow_memory_ViT import RM_ViT
-# from methods.ewc_ViT import EWCpp_ViT
-# from methods.mir_ViT import MIR_ViT
-from methods.clib_vit import CLIB_ViT
-
-logger = logging.getLogger()
-
-
-def select_method(args, criterion, device, train_transform, test_transform, n_classes):
-    kwargs = vars(args)
-    if args.mode == "er":
-        method = ER(
-            criterion=criterion,
-            device=device,
-            train_transform=train_transform,
-            test_transform=test_transform,
-            n_classes=n_classes,
-            **kwargs,
-        )
-    elif args.mode == "gdumb":
-        from methods.gdumb import GDumb
-        method = GDumb(
-            criterion=criterion,
-            device=device,
-            train_transform=train_transform,
-            test_transform=test_transform,
-            n_classes=n_classes,
-            **kwargs,
-        )
-    elif args.mode == "rm":
-        method = RM(
-            criterion=criterion,
-            device=device,
-            train_transform=train_transform,
-            test_transform=test_transform,
-            n_classes=n_classes,
-            **kwargs,
-        )
-    elif args.mode == "bic":
-        method = BiasCorrection(
-            criterion=criterion,
-            device=device,
-            train_transform=train_transform,
-            test_transform=test_transform,
-            n_classes=n_classes,
-            **kwargs,
-        )
-    elif args.mode == "ewc++":
-        method = EWCpp(
-            criterion=criterion,
-            device=device,
-            train_transform=train_transform,
-            test_transform=test_transform,
-            n_classes=n_classes,
-            **kwargs,
-        )
-    elif args.mode == "mir":
-        method = MIR(
-            criterion=criterion,
-            device=device,
-            train_transform=train_transform,
-            test_transform=test_transform,
-            n_classes=n_classes,
-            **kwargs,
-        )
-    elif args.mode == "clib":
-        method = CLIB(
-            criterion=criterion,
-            device=device,
-            train_transform=train_transform,
-            test_transform=test_transform,
-            n_classes=n_classes,
-            **kwargs,
-        )
-    elif args.mode == "clib_vit":
-        method = CLIB_ViT(
-            criterion=criterion,
-            device=device,
-            train_transform=train_transform,
-            test_transform=test_transform,
-            n_classes=n_classes,
-            **kwargs,
-        )
-    elif args.mode == "L2P":
-        method = L2P(
-            criterion=criterion,
-            device=device,
-            train_transform=train_transform,
-            test_transform=test_transform,
-            n_classes=n_classes,
-            **kwargs,
-        )
-    elif args.mode == "ViT_LP":
-        method = ViT_LP(
-            criterion=criterion,
-            device=device,
-            train_transform=train_transform,
-            test_transform=test_transform,
-            n_classes=n_classes,
-            **kwargs,
-        )
-    elif args.mode == "ViT_FT":
-        method = ViT_FT(
-            criterion=criterion,
-            device=device,
-            train_transform=train_transform,
-            test_transform=test_transform,
-            n_classes=n_classes,
-            **kwargs,
-        )
-    # elif args.mode == "Finetuning":
-    #     method = Finetuning(
-    #         criterion=criterion,
-    #         device=device,
-    #         train_transform=train_transform,
-    #         test_transform=test_transform,
-    #         n_classes=n_classes,
-    #         **kwargs,
-    #     )
-    # elif args.mode == "Freeze_extractor":
-    #     method = Freeze_extractor(
-    #         criterion=criterion,
-    #         device=device,
-    #         train_transform=train_transform,
-    #         test_transform=test_transform,
-    #         n_classes=n_classes,
-    #         **kwargs,
-    #     )
-    else:
-        raise NotImplementedError("Choose the args.mode in [er, gdumb, rm, bic, ewc++, mir, clib]")
-
-    return method
